Order with change status

+++ src/main/java/com/example/shopping_application/service/impl/OrderServiceImpl.java
@@ -46,32 +46,85 @@ public class OrderServiceImpl implements OrderService {
     public void removeByProductId(int id) {
         orderItemRepository.deleteByProduct_Id(id);
     }
-
     @Override
     @Transactional
     public void save(int userId) {
-        Optional<User> user = userRepository.findById(userId);
-        List<Cart> cartList = cartRepository.findAllByUserId(userId);
-        int totalAmount = 0;
-
-        Order order = new Order();
-        order.setUser(user.orElse(null));
+        Optional<User> userOptional = userRepository.findById(userId);
         List<OrderItem> orderItems = new ArrayList<>();
 
-        for (Cart cart : cartList) {
-            List<CartItem> cartItems = cart.getCartItems();
-            for (CartItem cartItem : cartItems) {
-                OrderItem orderItem = new OrderItem();
-                totalAmount += cartItem.getProduct().getPrice() * cartItem.getCount();
-                orderItem.setCount(cartItem.getCount());
-                orderItem.setProduct(cartItem.getProduct());
-                orderItems.add(orderItem);
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            List<Cart> cartList = cartRepository.findAllByUserId(userId);
+            int totalAmount = 0;
+            Order byUserIdAndStatus = orderRepository.findByUserIdAndStatus(userId, Status.PENDING);
+
+            if (byUserIdAndStatus == null) {
+                Order order = new Order();
+                order.setUser(user);
+
+                for (Cart cart : cartList) {
+                    List<CartItem> cartItems = cart.getCartItems();
+
+                    for (CartItem cartItem : cartItems) {
+                        boolean found = false;
+
+                        for (OrderItem orderItem : orderItems) {
+                            if (orderItem.getProduct().getId() == cartItem.getProduct().getId()) {
+                                orderItem.setCount(orderItem.getCount() + cartItem.getCount());
+                                found = true;
+                                break;
+                            }
+                        }
+
+                        if (!found) {
+                            OrderItem orderItem = new OrderItem();
+                            orderItem.setCount(cartItem.getCount());
+                            orderItem.setProduct(cartItem.getProduct());
+                            orderItems.add(orderItem);
+                        }
+
+                        totalAmount += cartItem.getProduct().getPrice() * cartItem.getCount();
+                    }
+                }
+
+                order.setOrderItems(orderItems);
+                order.setTotalAmount(totalAmount);
+                order.setStatus(Status.PENDING);
+                orderRepository.save(order);
+                cartRepository.deleteByUserId(userId);
+            } else {
+                Order existingOrder = byUserIdAndStatus;
+
+                for (Cart cart : cartList) {
+                    List<CartItem> cartItems = cart.getCartItems();
+
+                    for (CartItem cartItem : cartItems) {
+                        boolean found = false;
+
+                        for (OrderItem orderItem : existingOrder.getOrderItems()) {
+                            if (orderItem.getProduct().getId() == cartItem.getProduct().getId()) {
+                                orderItem.setCount(orderItem.getCount() + cartItem.getCount());
+                                found = true;
+                                break;
+                            }
+                        }
+
+                        if (!found) {
+                            OrderItem orderItem = new OrderItem();
+                            orderItem.setCount(cartItem.getCount());
+                            orderItem.setProduct(cartItem.getProduct());
+                            existingOrder.getOrderItems().add(orderItem);
+                        }
+
+                        totalAmount += cartItem.getProduct().getPrice() * cartItem.getCount();
+                    }
+                }
+
+                existingOrder.setTotalAmount(existingOrder.getTotalAmount() + totalAmount);
+                orderRepository.save(existingOrder);
+                cartRepository.deleteByUserId(userId);
             }
         }
-        order.setOrderItems(orderItems);
-        order.setTotalAmount(totalAmount);
-        orderRepository.save(order);
-        cartRepository.deleteByUserId(userId);
     }
 
 }
